[DUT-939]: shared와 packages 경계 정리
