+++ apps/app/src/features/auth/useRegister/index.ts
@@ -6,8 +6,8 @@ import {accountQueryOptions} from '@/entities/account/model/queries';
 import useLoadingUseCase from '@/features/ui/useLoading';
 import useTutorialUseCase from '@/features/ui/useTutorial';
 import {AccountAPI, NurseAPI, WardAPI} from '@/shared/api';
-import {type TCreateNurseDTO} from '@/shared/api/nurse/type';
-import {type TCreateWardDTO} from '@/shared/api/ward/type';
+import {type TCreateNurseDTO} from '@dutying/api/nurse';
+import {type TCreateWardDTO} from '@dutying/api/ward';
 import ROUTE from '@/shared/constant/path';
 import {showActionErrorFeedback} from '@/shared/util/feedback';
 import useAuth from '../useAuth';
