+++ apps/app/src/pages/onboarding-ward-create/model/adapter.ts
@@ -1,6 +1,6 @@
 import {v4 as uuidv4} from 'uuid';
 import {type TOnboardingWardParseApiResponse} from '@/shared/api/file/type';
-import {type TCreateWardDTO} from '@/shared/api/ward/type';
+import {type TCreateWardDTO} from '@dutying/api/ward';
 import {
     createEmptyShiftType,
     type TOnboardingNurseDraft,
