+++ apps/app/src/pages/onboarding-ward-create/model/draft.ts
@@ -1,5 +1,5 @@
 import {type DropResult} from '@hello-pangea/dnd';
-import {type TCreateWardDTO} from '@/shared/api/ward/type';
+import {type TCreateWardDTO} from '@dutying/api/ward';
 
 export type TOnboardingStep = 1 | 2 | 3 | 4;
