--- apps/app/src/shared/api/account/type.ts
@@ -1,1 +1,0 @@
-export type * from '@dutying/api/account';
