--- apps/app/src/shared/api/nurse/type.ts
@@ -1,1 +1,0 @@
-export type * from '@dutying/api/nurse';
