+++ apps/app/src/shared/api/ward/type.ts
@@ -1,13 +1,6 @@
-import type {IWardAPI as IBaseWardAPI} from '@dutying/api/ward';
-import type {TAiScheduleResponse} from '@/shared/types/ai-schedule';
+import type {IWardAPI as IBaseWardAPI, TAiScheduleResponse, TGenerateAiAutofillScheduleDTO} from '@dutying/api/ward';
 
 export type * from '@dutying/api/ward';
-
-export type TGenerateAiAutofillScheduleDTO = {
-    year: number;
-    month: number;
-    schedule: Record<string, string[]>;
-};
 
 export interface IWardAPI extends IBaseWardAPI {
     generateAiAutofillSchedule: (
