+++ packages/utils/package.json
@@ -19,6 +19,10 @@
         "./style": {
             "types": "./src/style.ts",
             "import": "./src/style.ts"
+        },
+        "./types": {
+            "types": "./src/types.ts",
+            "import": "./src/types.ts"
         }
     },
     "dependencies": {
