+++ packages/utils/src/index.ts
@@ -1,3 +1,4 @@
 export * from './date';
 export * from './ko-to-en';
 export * from './style';
+export * from './types';
